Fixed tests.
