+++ src/System.IO/tests/BinaryWriter/BinaryWriter.WriteByteCharTests.cs
@@ -9,7 +9,7 @@ using System.Text;
 
 namespace System.IO.Tests
 {
-    public class BinaryWriter_WriteByteCharTests
+    public partial class BinaryWriter_WriteByteCharTests
     {
         protected virtual Stream CreateStream()
         {
@@ -557,41 +557,6 @@ namespace System.IO.Tests
 
             mstr.Dispose();
             dw2.Dispose();
-        }
-
-        [Fact]
-        public void BinaryWriter_WriteSpan()
-        {
-            byte[] bytes = new byte[] { 4, 2, 7, 0xFF };
-            char[] chars = new char[] { 'a', '7', Char.MaxValue };
-            Span<byte> byteSpan = new Span<byte>(bytes);
-            Span<char> charSpan = new Span<char>(chars);
-
-            using (Stream memoryStream = CreateStream())
-            {
-                using (BinaryWriter binaryWriter = new BinaryWriter(memoryStream, Encoding.Unicode))
-                {
-                    binaryWriter.Write(byteSpan);
-                    binaryWriter.Write(charSpan);
-
-                    Stream baseStream = binaryWriter.BaseStream;
-                    baseStream.Position = 2;
-
-                    Assert.Equal(7, baseStream.ReadByte());
-                    Assert.Equal(0xFF, baseStream.ReadByte());
-
-                    char testChar;
-
-                    testChar = BitConverter.ToChar(new byte[] { (byte)baseStream.ReadByte(), (byte)baseStream.ReadByte() }, 0);
-                    Assert.Equal('a', testChar);
-
-                    testChar = BitConverter.ToChar(new byte[] { (byte)baseStream.ReadByte(), (byte)baseStream.ReadByte() }, 0);
-                    Assert.Equal('7', testChar);
-
-                    testChar = BitConverter.ToChar(new byte[] { (byte)baseStream.ReadByte(), (byte)baseStream.ReadByte() }, 0);
-                    Assert.Equal(Char.MaxValue, testChar);
-                }
-            }
         }
     }
 }
